feat(macos): keyboard capture tap — suppress system shortcuts, send KEY_EVENT (K-3) (#25)
* fix(protocol): KEY_EVENT metaState u16→u32 with real Android META_* constants

K-1 defined metaState with invented flags (0x1000=Shift etc.) that do not
match actual Android KeyEvent constants (META_SHIFT_ON=0x1, META_CTRL_ON=
0x1000) and could not represent META_META_ON=0x10000 in u16. The Android
helper (K-2) already uses real Android constants, so the wire definition
would have misrouted Shift as Ctrl. Align the spec, fixtures, and both
implementations on the real constants; metaState is now u32.

- protocol.md: KEY_EVENT payload u32 + real META_* values
- key-event fixture: metaState 0x1000 (was 'Shift') → 0x1 (real Shift)
- Swift Protocol.swift: builder/decoder u32; tests updated (21/21 green)
- Kotlin Messages.kt: parser u32; CodecTest updated (all suites green)
- check-fixtures.mjs: readUInt32LE (13 fixtures verified)

Refs #21

* feat(macos): keyboard capture tap — suppress system shortcuts, send KEY_EVENT (ADR-0007)

K-3: the macOS capture side of keyboard delivery.

- InputCapture: captures keyDown/keyUp/flagsChanged. While suppressed, keyboard
  events are consumed so Cmd+Tab / Spotlight / Mission Control never fire on
  macOS — the previously-blocking open problem. Forwarded as Android key-code
  transitions with real Android META_* modifiers: onKeyEvent callback.
- Fail-safe (AGENTS.md rule 5): keys held when suppression ends (timeout,
  disconnect, emergency ⌘⇧X) are flushed as UP events so the device never keeps
  a stuck key.
- KeyCodeMapper: macOS kVK_* → Android KEYCODE_* + CGEventFlags → META_* bits,
  pure and unit-tested; non-ANSI keys (media/brightness) are consumed but not
  forwarded.
- App: onCapturedKeyEvent sends the KEY_EVENT frame; helper picks the backend.

Verification: swift build OK; swift test 27/27 (KeyCodeMapper 7 new tests).

Refs #21

+++ apps/macos/Sources/Protocol/Protocol.swift
@@ -176,8 +176,8 @@ public enum Messages {
 
     /// Android KeyEvent semantics: keyCode = KeyEvent.KEYCODE_*, metaState = KeyEvent.META_*.
     /// action: 0=KEY_ACTION_DOWN, 1=KEY_ACTION_UP. repeatCount: 0 = first press.
-    public static func keyEvent(keyCode: UInt16, metaState: UInt16, action: UInt8, repeatCount: UInt8 = 0) -> Data {
-        LE.u16(keyCode) + LE.u16(metaState) + LE.u8(action) + LE.u8(repeatCount)
+    public static func keyEvent(keyCode: UInt16, metaState: UInt32, action: UInt8, repeatCount: UInt8 = 0) -> Data {
+        LE.u16(keyCode) + LE.u32(metaState) + LE.u8(action) + LE.u8(repeatCount)
     }
 }
 
@@ -262,10 +262,10 @@ public extension Messages {
         return try d.u16()
     }
 
-    static func decodeKeyEvent(_ payload: Data) throws -> (keyCode: UInt16, metaState: UInt16, action: UInt8, repeatCount: UInt8) {
+    static func decodeKeyEvent(_ payload: Data) throws -> (keyCode: UInt16, metaState: UInt32, action: UInt8, repeatCount: UInt8) {
         var d = Decoder(payload)
         let keyCode = try d.u16()
-        let metaState = try d.u16()
+        let metaState = try d.u32()
         let action = try d.u8()
         let repeatCount = try d.u8()
         return (keyCode, metaState, action, repeatCount)
